ajuste

+++ src/main/resources/db/migration/V1__criar_tabela_pais.sql
@@ -1,6 +1,6 @@
 CREATE TABLE tb_pais
 (
-    codigo bigint       PRIMARY KEY,
+    codigo serial    PRIMARY KEY,
     nome   varchar(100) NOT NULL,
     PRIMARY KEY (codigo)
 ) ENGINE = InnoDB
